+++ portlets/lecture2go-portlet/docroot/WEB-INF/service/de/uhh/l2g/plugins/service/ProducerLocalServiceClp.java
@@ -742,7 +742,7 @@ public class ProducerLocalServiceClp implements ProducerLocalService {
 	}
 
 	@Override
-	public java.util.List<java.lang.Integer> getAllProducerIds(
+	public java.util.List<java.lang.Long> getAllProducerIds(
 		java.lang.Long lectureseriesId)
 		throws com.liferay.portal.kernel.exception.SystemException {
 		Object returnObj = null;
@@ -768,11 +768,11 @@ public class ProducerLocalServiceClp implements ProducerLocalService {
 			}
 		}
 
-		return (java.util.List<java.lang.Integer>)ClpSerializer.translateOutput(returnObj);
+		return (java.util.List<java.lang.Long>)ClpSerializer.translateOutput(returnObj);
 	}
 
 	@Override
-	public java.util.List<java.lang.Integer> getProducerIds(
+	public java.util.List<java.lang.Long> getProducerIds(
 		java.lang.Long lectureseriesId, int begin, int end)
 		throws com.liferay.portal.kernel.exception.SystemException {
 		Object returnObj = null;
@@ -804,7 +804,7 @@ public class ProducerLocalServiceClp implements ProducerLocalService {
 			}
 		}
 
-		return (java.util.List<java.lang.Integer>)ClpSerializer.translateOutput(returnObj);
+		return (java.util.List<java.lang.Long>)ClpSerializer.translateOutput(returnObj);
 	}
 
 	private InvokableLocalService _invokableLocalService;
